refactor: extract dimensionChange functionality to container
